+++ projects/ngx-webrtc-lib/src/lib/services/video-call-dialog.service.ts
@@ -7,6 +7,11 @@ import { VideoCallDialog, VideoCallDialogData } from '../models';
 import { WebRtcComponent } from '../webrtc.component';
 import { DialogService } from './dialog.service';
 
+/**
+ * Service for managing the video call dialog and its interactions.
+ * Provides methods to open a video call dialog and handle call flow.
+ * @publicApi
+ */
 @Injectable({
   providedIn: 'root'
 })
@@ -14,6 +19,13 @@ export class VideoCallDialogService {
 
   constructor(private dialog: DialogService) { }
 
+  /**
+   * Opens the video call dialog with the given data.
+   * Manages the flow of the video call including confirming the call and launching WebRTC for video.
+   *
+   * @param data The data to initialize the video call dialog (includes user info, channel, etc.).
+   * @returns An object with methods to control the dialog and get call status.
+   */
   open(data: VideoCallDialogData): VideoCallDialog {
     const callEnded = new Subject<boolean>();
     const dialog = this.dialog.open(VideoCallComponent, { data }) as VideoCallComponent;
@@ -25,7 +37,6 @@ export class VideoCallDialogService {
             uid: data.uid,
             token: data.token,
             channel: data.channel,
-            debug: !!data.debug,
             displaySmallScreen: true,
           }).afterClosed.asObservable();
         }